requirements.txt
python-dotenv~=1.0.1
Django~=5.0.7
usos-api~=0.2.0
requests~=2.32.3